Changes to IsolateCourseOnDayConstriction.java, Added UnbalancedDaysPenalization
IsolateCourseOnDayConstriction.java now can be relaxed. In case that because of it there are no viable days, then the conditions will be relaxed so that there must be no exam of the same course and semester on the same day.

UnbalancedDaysPenalization.java, is a new type of weak constriction aimed to penalize unbalanced timetables, however it makes no real effect. Needs a redesign.